Restructured files.
