+++ Chess Project/src/chess/project/players/AIPlayer.java
@@ -1,15 +1,19 @@
-package chess.project;
+package chess.project.players;
+
+import chess.project.Move;
 
 /**
  *
  * @author Alex Zurad
  */
-public class HumanPlayer extends Player{
+public class AIPlayer extends Player{
 
     @Override
     public Move getMove() {
         
         return null;
     }
+
+
     
 }
