+++ Chess Project/src/chess/project/players/HumanPlayer.java
@@ -1,17 +1,17 @@
-package chess.project;
+package chess.project.players;
+
+import chess.project.Move;
 
 /**
  *
  * @author Alex Zurad
  */
-public class AIPlayer extends Player{
+public class HumanPlayer extends Player{
 
     @Override
     public Move getMove() {
         
         return null;
     }
-
-
     
 }
